Remove PopupViewTestViewModel
